Update Admin DB

+++ app/Http/Controllers/Admin/Pages/MajorsController.php
@@ -11,81 +11,84 @@ use Illuminate\View\View;
 
 class MajorsController extends Controller
 {
-    public function index() : View
+    public function index(): View
     {
-        $majors  = Major::get(); 
+        $majors  = Major::get();
 
-        return view('admin.pages.major.index',compact('majors'));
-    } 
-    public function create(): View 
-    {    
+        return view('admin.pages.major.index', compact('majors'));
+    }
+    public function create(): View
+    {
         return view('admin.pages.major.create');
     }
-    public function store(MajorRequest $request): RedirectResponse 
+    public function store(MajorRequest $request): RedirectResponse
     {
         try {
 
-            Major::create([   
-                'name' => [
-                    'en' => $request->input("name.en"),
-                    'ar' => $request->input("name.ar")
-                    ],
-            ]);
-            
-            return redirect()->route('admins.majors.index')->with("success","major created successfully");
-            
-        } catch (Exception $e) {
-            dd($e->getMessage());
-            return redirect()->back();//message
-        }
-    }
-
-    public function edit(Major $major) : View
-    { 
-        return view('admin.pages.major.edit',compact('major'));
-    }
-    
-    public function update(MajorRequest $request, Major $major) : RedirectResponse
-    {
-        try {
-
-            $major->update([   
+            Major::create([
                 'name' => [
                     'en' => $request->input("name.en"),
                     'ar' => $request->input("name.ar")
                 ],
             ]);
 
-            return redirect()->route('admins.majors.index')->with("success","major updated successfully");
+            return redirect()->route('admins.majors.index')->with("success", "major created successfully");
         } catch (Exception $e) {
-            dd($e->getMessage());
-            return redirect()->back();//message
+            // dd($e->getMessage());
+            return redirect()->back()->with('error',$e->getMessage());//message
         }
     }
-    
-    public function destroy(Major $major) : RedirectResponse 
+
+    public function edit(Major $major): View
+    {
+        return view('admin.pages.major.edit', compact('major'));
+    }
+
+    public function update(MajorRequest $request, Major $major): RedirectResponse
     {
         try {
+
+            $major->update([
+                'name' => [
+                    'en' => $request->input("name.en"),
+                    'ar' => $request->input("name.ar")
+                ],
+            ]);
+
+            return redirect()->route('admins.majors.index')->with("success", "major updated successfully");
+        } catch (Exception $e) {
+            // dd($e->getMessage());
+            return redirect()->back()->with('error',$e->getMessage());//message
+        }
+    }
+
+    public function destroy(Major $major): RedirectResponse
+    {
+        try {
+            if ($major->serviceProviders->count() > 0) {
+                return redirect()->back()->with('error','can\'t delete, there are active service providers in this major');   
+            }
+
             $major->delete();
 
-            return redirect()->route('admins.majors.index');
+            return redirect()->route('admins.majors.index')->with('success', 'major deleted successfully');
         } catch (Exception $e) {
-            dd($e->getMessage());
-            return redirect()->back();//message
+            // dd($e->getMessage());
+            return redirect()->back()->with('error',$e->getMessage());//message
         }
     }
 
-    public function stauts(Major $major) : RedirectResponse 
+    public function stauts(Major $major): RedirectResponse
     {
         try {
-            $major->update([   
+            $major->update([
                 'active' => !$major->active
             ]);
 
             return redirect()->route('admins.majors.index');
         } catch (Exception $e) {
             dd($e->getMessage());
-            return redirect()->back();//message
+            return redirect()->back(); //message
         }
     }
 }
